Renamed the project to JobBook from EntAppProject. Renamed welcome screen to index.html
