+++ EntAppProject/src/main/java/com/jobook/model/Job.java
@@ -1,4 +1,4 @@
-package com.jobook.EntAppProject.model;
+package com.jobook.model;
 
 import java.util.Date;
