+++ EntAppProject/src/main/java/com/jobook/model/Skill.java
@@ -1,4 +1,4 @@
-package com.jobook.EntAppProject.model;
+package com.jobook.model;
 
 public class Skill {
 	Integer id;
