+++ EntAppProject/src/test/java/com/jobook/EntAppProject/UserServiceUnitTests.java
@@ -11,10 +11,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import com.jobook.EntAppProject.Services.IUserService;
-import com.jobook.EntAppProject.model.Customer;
-import com.jobook.EntAppProject.model.Employee;
-import com.jobook.EntAppProject.model.User;
+import com.jobook.Services.IUserService;
+import com.jobook.model.Customer;
+import com.jobook.model.Employee;
+import com.jobook.model.User;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
